# docs/API.yaml
openapi: 3.0.0
info:
  title: UdaConnect Persons and Person Interconnections API
  version: 0.1.0
  contact:
    name: Karan Sharad Mistry
    email: karanmistry284@gmail.com
...
paths:
...
  "/api/persons/{person_id}/connection":
    servers:
      - url: http://localhost:30002/
    parameters:
      - in: path
        description: Unique ID for a given Person
        name: person_id
        required: true
        schema:
          type: string
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionSchema'
      operationId: get_person_connections_resource
      tags:
        - UdaConnect
      summary: Retrieves specified person
      description: Retrieves a json formatted person data by specifying its ID
components:
  schemas:
    PersonSchema:
      type: object
      required:
        - first_name
        - last_name
        - company_name
      properties:
        id:
          type: integer
          format: int32
          example: 10
        first_name:
          type: string
          example: David
        last_name:
          type: string
          example: Johnson
        company_name:
          type: string
          example: P&G
    LocationSchema:
      type: object
      required:
        - person_id
        - longitude
        - latitude
        - creation_time
      properties:
        id:
          type: integer
          format: int32
          example: 45
        person_id:
          type: integer
          format: int32
          example: 11
        longitude:
          type: string
          example: -182.45083
        latitude:
          type: string
          example: 93.34267
        creation_time: 
          type: string
          example: "2021-08-05T10:37:06"
    ConnectionSchema:
      type: array
      items:
        properties:
          person:
            $ref: '#/components/schemas/PersonSchema'
          location:
            $ref: '#/components/schemas/LocationSchema'